used tanstack query and used refetch where needed

+++ src/pages/Dashboard/SelectedCourses/CourseRow/CourseRow.jsx
@@ -1,21 +1,45 @@
+import Swal from "sweetalert2";
+import useSelected from "../../../../hooks/useSelected";
 
 
-const CourseRow = ({course, index}) => {
+const CourseRow = ({ course, index }) => {
+
+    // to refetch in useSelected hook
+    const [, refetch, ,] = useSelected();
 
 
     const handleDelete = () => {
-        fetch(`http://localhost:5000/selectedCourses/${course._id}`, {
-            method: 'DELETE'
-        })
-        .then(res=>res.json())
-        .then(data=>{
-            if(data.deletedCount>0) {
-                alert('deleted')
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: 'rgb(220 38 38)',
+            cancelButtonColor: '#636e72',
+            confirmButtonText: 'Yes, delete the Course'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/selectedCourses/${course._id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            refetch();
+                            Swal.fire(
+                                'Deleted!',
+                                'The course has been deleted.',
+                                'success'
+                            )
+                        }
+                    })
+                    .catch(err => {
+                        console.log(err.message);
+                    })
             }
         })
-        .catch(err=>{
-            console.log(err.message);
-        })
+
     }
 
     return (
